requirements.txt
# Core ML dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0

# Whisper specific
openai-whisper>=20231117
whisper-normalizer>=0.1.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0
pydub>=0.25.0

# Text-to-Speech for data generation
gTTS>=2.4.0
edge-tts>=6.1.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
jiwer>=3.0.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Visualization and logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
wandb>=0.16.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
rich>=13.0.0
click>=8.1.0

# Optional: For better performance
# flash-attn>=2.3.0  # Uncomment if you have compatible GPU
